ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pyannote.audio==3.1.1
torch==2.1.0
torchaudio==2.1.0
pymorphy3==2.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
